ATC-273 | cli: extract client construction, project resolution, and attach into internal/cli
