+++ cmd/atc/project.go
@@ -1,19 +1,15 @@
 package main
 
 import (
-	"bufio"
 	"fmt"
 	"io"
-	"os"
-	"os/exec"
 	"path/filepath"
-	"strings"
 	"text/tabwriter"
 
 	"github.com/spf13/cobra"
 
 	"github.com/jeremytondo/atc/internal/api"
-	"github.com/jeremytondo/atc/internal/paths"
+	"github.com/jeremytondo/atc/internal/cli"
 )
 
 func newProjectCmd() *cobra.Command {
@@ -55,7 +51,7 @@ directory itself must exist on the server's machine.`,
 				if directory, err = filepath.Abs(args[0]); err != nil {
 					return err
 				}
-			} else if directory, err = defaultProjectDir(); err != nil {
+			} else if directory, err = cli.DefaultProjectDir(); err != nil {
 				return err
 			}
 			project, err := client.CreateProject(cmd.Context(), api.ProjectCreateParams{
@@ -151,93 +147,6 @@ those first; there is no cascade.`,
 			return err
 		}),
 	}
-}
-
-// stdinIsTTY gates the create-offer prompt: a question is only asked of a
-// human on an interactive stdin. Stdout deliberately does not matter — a
-// redirected stdout still deserves the prompt (on stderr) rather than a
-// refusal. A variable so tests, which never have a TTY, can force it.
-var stdinIsTTY = func() bool {
-	info, err := os.Stdin.Stat()
-	return err == nil && info.Mode()&os.ModeCharDevice != 0
-}
-
-// resolveProjectID finds the project owning the current directory the way
-// git finds a repository (ATC-256): canonicalize the cwd, then walk up —
-// the nearest ancestor (or the directory itself) matching a project's
-// directory wins. On no match, an interactive run is offered a project
-// rooted at the default (git toplevel, else cwd); a non-interactive run
-// refuses with the fixes instead of prompting.
-func resolveProjectID(cmd *cobra.Command, client *api.Client) (string, error) {
-	canonical, err := paths.CanonicalDir(".")
-	if err != nil {
-		return "", err
-	}
-	projects, err := client.Projects(cmd.Context())
-	if err != nil {
-		return "", err
-	}
-	byDir := make(map[string]string, len(projects))
-	for _, project := range projects {
-		byDir[project.Directory] = project.ID
-	}
-	for dir := canonical; ; {
-		if id, ok := byDir[dir]; ok {
-			return id, nil
-		}
-		parent := filepath.Dir(dir)
-		if parent == dir {
-			break
-		}
-		dir = parent
-	}
-
-	if !stdinIsTTY() {
-		return "", fmt.Errorf("no project contains %s; pass --project <id> or run `atc project create`", canonical)
-	}
-	defaultDir := projectRootFor(canonical)
-	// The conversation goes to stderr: stdout stays the command's
-	// redirectable output (the created terminal).
-	console := cmd.ErrOrStderr()
-	_, _ = fmt.Fprintf(console, "no project contains %s\ncreate a project at %s? [y/N] ", canonical, defaultDir)
-	answer, _ := bufio.NewReader(cmd.InOrStdin()).ReadString('\n')
-	switch strings.ToLower(strings.TrimSpace(answer)) {
-	case "y", "yes":
-	default:
-		return "", fmt.Errorf("no project selected; pass --project <id> or run `atc project create`")
-	}
-	project, err := client.CreateProject(cmd.Context(), api.ProjectCreateParams{Directory: defaultDir})
-	if err != nil {
-		return "", err
-	}
-	_, _ = fmt.Fprintf(console, "created project %s at %s\n", project.ID, project.Directory)
-	return project.ID, nil
-}
-
-// defaultProjectDir is where an unspecified project is rooted: the git
-// toplevel when inside a repository, the current directory otherwise.
-func defaultProjectDir() (string, error) {
-	canonical, err := paths.CanonicalDir(".")
-	if err != nil {
-		return "", err
-	}
-	return projectRootFor(canonical), nil
-}
-
-// projectRootFor asks git for the toplevel of the repository containing
-// the canonical directory — git itself is the authority on what counts as
-// a repository (a stray or malformed .git entry is not one). Outside a
-// repository, or without git installed, the directory is its own root.
-func projectRootFor(canonical string) string {
-	output, err := exec.Command("git", "-C", canonical, "rev-parse", "--show-toplevel").Output()
-	if err != nil {
-		return canonical
-	}
-	toplevel, err := paths.CanonicalDir(strings.TrimSpace(string(output)))
-	if err != nil {
-		return canonical
-	}
-	return toplevel
 }
 
 func printProject(out io.Writer, project api.Project) {
