+++ cmd/atc/project_test.go
@@ -7,6 +7,7 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/jeremytondo/atc/internal/cli"
 	"github.com/jeremytondo/atc/internal/paths"
 )
 
@@ -247,7 +248,7 @@ func TestProjectCreateDefaultPath(t *testing.T) {
 	plain := canonical(t, t.TempDir())
 	// The walk honestly finds any ancestor repo (a stray /tmp/.git on a
 	// developer machine, say); only a truly repo-free cwd tests this case.
-	if projectRootFor(plain) != plain {
+	if cli.ProjectRootFor(plain) != plain {
 		t.Skipf("an ancestor of %s is a git repository; skipping the no-repo case", plain)
 	}
 	t.Chdir(plain)
