+++ cmd/atc/terminal_test.go
@@ -13,6 +13,7 @@ import (
 	"sync"
 	"testing"
 
+	"github.com/jeremytondo/atc/internal/cli"
 	"github.com/jeremytondo/atc/internal/events"
 	"github.com/jeremytondo/atc/internal/projects"
 	"github.com/jeremytondo/atc/internal/server"
@@ -122,10 +123,10 @@ func createProjectCLI(t *testing.T, dir string) string {
 
 func forceTTY(t *testing.T) {
 	t.Helper()
-	prevStdio, prevStdin := stdioIsTerminal, stdinIsTTY
-	stdioIsTerminal = func() bool { return true }
-	stdinIsTTY = func() bool { return true }
-	t.Cleanup(func() { stdioIsTerminal, stdinIsTTY = prevStdio, prevStdin })
+	prevStdio, prevStdin := cli.StdioIsTerminal, cli.StdinIsTTY
+	cli.StdioIsTerminal = func() bool { return true }
+	cli.StdinIsTTY = func() bool { return true }
+	t.Cleanup(func() { cli.StdioIsTerminal, cli.StdinIsTTY = prevStdio, prevStdin })
 }
 
 func installFakeZmx(t *testing.T) {
